Swap to assuming no .pyc for build descriptions
(this necessitated changing the test scripts to make the .pyc optional)

Not producing the .pyc file is good because it means that a change
to the .py file will always take affect (the test scripts had to
delete the .pyc file because of Python's wish to read the .pyc
if it is "up to date", and not being very good at detecting that),
but potentially bad because loading a .py file is slightly slower
than loading a .pyc file.

However, if we're going to want to detect changes to the build
description, and we will if we're going to cache our dependency
trees in a database, then losing the .pyc files simplifies things.