Converted tag storage from file/folder based to sqlite3 databases
Each .muddle folder usually contains its own tags folder containing many files just
containing time stamps, the sqlite database is now in the same location as the tags folder would
be (subdomains contain their own databases currently). An "in progress" status
was added in addition to the current done or undone states.

This is currently not backwards compatible (tag folders are not ported over to a database)
but it passes all tests.