Parallelised muddle for basic cases including breaking config changes
Parallelised muddle based upon running multiple processes. For testing
purposes this has been automated so that the normal
builder.build_label(s) will spawn subprocesses in their own terminals so
that the test scripts can be used to test running many muddles in
parallel. The terminals will currently exit unless the muddle subprocess
returns an error code.

Tests that just build things generally work (though some are broken or
less capable as they haven't been adapted to read the new tag format)
and many of the other tests are currently broken. Domains are handled
mostly correctly for normal build tasks but are broken for at least
unstamping.

Dependencies are generally working correctly though being handled
inefficiently.

The system for storing tags is incompatible with the previous system of
files and will not currently convert existing tags to the new format, so
updating an existing muddle build tree is only possible by deleting the
currently existing checked out files or generating a new muddle build
tree.

The database schema seems to work reasonably (though dumping pickles of
rules is a bit hacky) but its usage will likely be changed so that the
current tables detail the results of the build description and will only
be updated when a build description changes, with another table(s) for
referencing labels/rules that need to be built for the current command.

Currently a secondary process to act on the same build command may be
started by running:

muddle _sub_builder_db

in the folder that the primary command was started in. For now there is
also the option of changing num_procs value in the Builder
initialisation in muddled/mechanics.py to automatically start
subprocesses.

The current mechanism of a master process and pausing other processes is
mostly unnecessary and was implemented for when all of the processes
wrote to the same stdout and read from the same stdin.