Changed database to hold all of ruleset rather than build specific subsets
Previously the database marked a rule as to be built by including it in
the database, which prevented any saved rule state from being used in
subsequent runs. This has now been changed to having all of the ruleset
state stored in the database with a separate table holding the targets
of the rules to be built.

This has also resulted in a decrease in startup time for large rulesets
as the previous strategy made use of recursion to find the necessary
labels and rules for a given target. The startup is still slower than
with base muddle, just not as horrifically slow.

Currently the ruleset tables are cleared and rewritten each time, the
eventual goal is to have keep this stored and only clearing and updating
it when a build description script changes, detected by a stored hash of
the file not matching.