Converted outputs to use the logging framework
The conversion was reasonably crude, due to the vast number of print
statements throughout the project they were first converted to print function
calls using 2to3 and then a print function was defined and an IDE tool was
used to change the name of the function. So far all of the keyword arguments
are ignored and it would probably be easier to change all of the places that
they are used than to add support for them.