Removed pausing mechanism and further converted logging
The logging mechanism was previously in place to allow for multiple processes
to print to a single terminal with sudo still being possible by waiting for
other threads to stop and then requesting the password. This is unnecessary
with the current structure of calling each sub-process in a separate console so
the mechanism can be removed.

Some files were outputting directly to sys.stdout rather than using print
statements and these have now been converted to using the logging framework.