Fixed muddle distribute and assorted other fixes
Fixed muddle distribute so that it will work with mutliple processes, though
the system to do so is a bit hacky. The distribute rules are transient but not
idempotent so the current system sets them all to be non-transient before
building and clears their status as the command finishes. Ideally there should
be a different system to distinguish between session transient labels which
should be run once each session and process transient labels that should be
run once for each process.

The database code has been changed to prevent acquiring a connection to a
database when the current process already has one, this prevents locking
errors where a database write is waiting for a read lock to be released
further up the stack.

A facility has been added to the logging module to allow for more detailed
output from a subset of modules rather than setting the logging level
universally.