Added retrying to deregister proc. and sub builders
Changed the code to deregister processes on build finishing or crashing so
that it would retry until it was successful as leaving the database with
processes that don't exist can cause problems with other areas. Code was also
added to release processing status on any rules that the process holds to
prevent a rule from being held in a non-working state.

The code for building was also changed so that sub-builders would wait for
dependencies to be satisfied if there are only unsatisfied rules left. The
previous behaviour was to terminate when no rule was immediately available to
build which wasn't helpful for unsupervised operation, or supervised operation
for that matter.

The logging was also changed minorly. The largest changes were changing the
limit to the number of log files from 50 to 12 and changing the logging
level to write to file to INFO from DEBUG, both measures to prevent the log
files from taking up large amounts of space.